moved add input/button to own component to reduce rerenders of list
